'add item' added

+++ app/src/main/java/com/example/lostandfound/DatabaseAndTables/User.java
@@ -2,7 +2,7 @@ package com.example.lostandfound.DatabaseAndTables;
 
 public class User {
 
-    int id;
+    long id;
     String name;
     String email;
     String phone;
@@ -12,7 +12,7 @@ public class User {
 
     }
 
-    public User(int id, String name, String email, String phone, String password) {
+    public User(long id, String name, String email, String phone, String password) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -20,11 +20,11 @@ public class User {
         this.password = password;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
